+++ satpy_overlay_plots.py
@@ -34,58 +34,12 @@ import botocore
 import fnmatch
 
 from oco_vistool import read_shp
+from geo_scan_time import compute_time_offset, approx_scanline_time
+from geo_scan_time import get_ABI_timerange_from_filename, get_AHI_timerange_from_filename
 
 # to get the location of the city data (a subdir from where the
 # code is located).
 _code_dir = os.path.dirname(os.path.realpath(__file__))
-
-def _approx_scanline_time(stime, etime, lat, domain):
-    """
-    Returns the scanline time, for a particular lat, and domain (C or F,
-    for (C)ONUS or (F)ull disk.)
-    see docstring in get_ABI_files for implementation notes.
-
-    Note this is a rough approximation, assuming the actual collection
-    time for a given scan line is a linear interpolation between the
-    start and end time in the file. It also assumes the north-most
-    scan line is collected first (Should be true for ABI, AHI).
-    This linear interpolation is not exactly true, because the actual
-    scan pattern is much more complicated. Doing the correct
-    calculation would be very difficult, and this approximation should
-    be accurate to a minute or two.
-
-    returns a single datetime object with the approx scanline time.
-
-    if the lat is out of range, it returns the closest time; in other
-    words, out of range will be limited to pick the stime or etime.
-
-    inputs:
-    stime: datetime object with start time
-    etime: datetime object with end time
-    lat: latitude (scalar)
-    domain: (C)ONUS or (F)ull disk
-    """
-
-    if domain == 'C':
-        slat = 50.0
-        elat = 15.0
-    elif domain == 'F':
-        slat = 81.0
-        elat = -81.0
-    else:
-        raise ValueError('invalid domain: '+domain)
-
-    tfrac = (slat - lat) / (slat - elat)
-
-    if tfrac < 0:
-        scan_time = stime
-    elif tfrac > 1:
-        scan_time = etime
-    else:
-        total_time = (etime-stime).total_seconds()
-        scan_time = stime + datetime.timedelta(seconds=total_time*tfrac)
-
-    return scan_time
 
 def _get_AHI_times(file_list):
     """
@@ -101,36 +55,6 @@ def _get_AHI_times(file_list):
     time_range = scn.start_time, scn.end_time
 
     return time_range
-
-
-def compute_time_offset(scn, lat, domain, datetime_utc):
-    """
-    compute the time offset from a given datetime in UTC, relative
-    to the observation time of a geostationary image.
-
-    uses _get_scanline_time() to estimate the actual observation time
-    of the time in the image.
-
-    Inputs:
-    scn : a satpy Scene object containing the geostationary image data
-    lat : the center lat of the requested region
-    domain : (C)ONUS or (F)ull disk
-    datetime_utc : a python datetime object, assumed in UTC, for the
-        requested time
-
-    returns:
-    time_offset : the offset between the input time and the observation time.
-        In other words, datetime_utc - obs_time; reported in seconds.
-    """
-
-    stime = scn.start_time
-    etime = scn.end_time
-
-    scan_time = _approx_scanline_time(stime, etime, lat, domain)
-
-    time_offset = (datetime_utc - scan_time).total_seconds()
-
-    return time_offset
 
 
 def get_loc_ABI_files(datetime_utc, data_home, domain, platform, hour_offsets, band_list, glob_fstr, verbose):
@@ -321,11 +245,8 @@ def get_ABI_files(datetime_utc, center_lat,
         time_offsets_all = []
         for n in range(nfiles):
             # find the approx scan line time for the requested latitude.
-            stimestamp, etimestamp = flists[band][n].split('_')[-3:-1]
-            # [1:-1] slice removes 's' or 'e' and the extra number (frac seconds?)
-            stime = datetime.datetime.strptime(stimestamp[1:-1], '%Y%j%H%M%S')
-            etime = datetime.datetime.strptime(etimestamp[1:-1], '%Y%j%H%M%S')
-            scan_time = _approx_scanline_time(stime, etime, center_lat, domain)
+            etime, stime = get_ABI_timerange_from_filename(flists[band][n])
+            scan_time = approx_scanline_time(stime, etime, center_lat, domain)
             time_offsets_all.append((datetime_utc-scan_time).total_seconds())
         # find minimum time difference
         k = np.argmin(np.abs(time_offsets_all))
@@ -478,10 +399,8 @@ def get_AHI_files(datetime_utc, center_lat, files_loc, data_home):
             # this means we don't have to download the file data.
             # the more accurate variation would use the _get_AHI_times() helper
             # function to get the actual start and end time.
-            segment_filename = files[band][n][0].split('/')[-1]
-            stime = datetime.datetime.strptime(segment_filename[7:20], '%Y%m%d_%H%M')
-            etime = stime + datetime.timedelta(minutes=10)
-            scan_time = _approx_scanline_time(stime, etime, center_lat, 'F')
+            stime, etime = get_AHI_timerange_from_filename(files[band][n][0])
+            scan_time = approx_scanline_time(stime, etime, center_lat, 'F')
             time_offsets_all.append((datetime_utc-scan_time).total_seconds())
         # find minimum time difference
         k = np.argmin(np.abs(time_offsets_all))
